# student-dropout-prediction/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Student Dropout Early-Warning System API
  description: |
    Role-Based Backend API for the Student Dropout Early-Warning System.
    Provides endpoints for student risk assessment, mentor intervention tracking,
    cohort reporting, bias & privacy auditing, and mentor management.
    
    ### Roles & Access Levels:
    - **Admin**: Full access across all departments, mentors, audit logs, fairness evaluation, and mentor user management.
    - **Mentor**: Scoped access limited to students assigned to the authenticated mentor. Audit and mentor management endpoints are restricted.
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Local Development FastAPI Server
security:
  - BearerAuth: []
...
paths:
...
  /api/mentors:
    get:
      tags:
        - Mentor Management
      summary: List all mentors (Admin Only)
      description: Returns all mentor accounts, status, and assigned student counts.
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MentorProfile'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenAdminOnlyError'
    post:
      tags:
        - Mentor Management
      summary: Register a new mentor account (Admin Only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MentorCreateRequest'
      responses:
        '201':
          description: Mentor created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MentorProfile'
        '400':
          description: Username already exists or validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenAdminOnlyError'
  /api/mentors/{id}:
    patch:
      tags:
        - Mentor Management
      summary: Edit mentor profile details (Admin Only)
      description: Updates name or email. Mentor ID and username cannot be changed.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MentorProfile'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenAdminOnlyError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /api/mentors/{id}/deactivate:
    patch:
      tags:
        - Mentor Management
      summary: Soft-deactivate or reactivate mentor (Admin Only)
      description: Toggles mentor active status. Deactivation prevents login while preserving assigned student data.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Status toggled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MentorProfile'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenAdminOnlyError'
        '404':
          $ref: '#/components/responses/NotFoundError'

# ==========================================
# COMPONENTS & SCHEMAS
# ==========================================
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    UnauthorizedError:
      description: Authentication required or invalid/expired session
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ForbiddenAdminOnlyError:
      description: Restricted endpoint - Admin role required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    MentorProfile:
      type: object
      properties:
        id:
          type: string
          example: mentor1
        username:
          type: string
          example: priya
        name:
          type: string
          example: Dr. Priya Nair
        email:
          type: string
          example: priya@university.edu
        role:
          type: string
          example: Mentor
        mentorId:
          type: string
          example: M001
        status:
          type: string
          enum: [Active, Inactive]
          example: Active
        assigned_students_count:
          type: integer
          example: 4
    MentorCreateRequest:
      type: object
      required:
        - name
        - username
        - password
      properties:
        name:
          type: string
          example: Dr. Ramesh Gupta
        username:
          type: string
          example: ramesh.gupta
        password:
          type: string
          example: Mentor@2026
        mentorId:
          type: string
          example: M004
        email:
          type: string
          example: ramesh@university.edu
    ErrorResponse:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
          example: Invalid credentials or access restricted.